# Fix perf regressions in Utf8Formatter for integers