@DrewScoggins 2023-08-24
Still seeing a regression in this test https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/main_x64_ubuntu%2018.04/System.Buffers.Text.Tests.Utf8FormatterTests.FormatterDecimal(value%3a%20123456.789).html

Found during 7.0 -> 8.0 perf comparison.